Moved AI to Sophie-AI-Model
This should make it easier to keep track of Sophie's AI Model, and a softlink was provided to link back one directory and to the area where the AI model is expected.

If you are a developer: Be sure to clone both sophiebot and Sophie-AI-Model and keep both repos in the same parent folder.